Production/cloud-computing/backend-talas/cloudfun/requirements.txt
pymysql
requests
google-cloud-firestore
cryptography